requirements.txt
fastapi==0.110.2
pika==1.3.2
uvicorn[standard]==0.29.0
pydantic[email]==2.7.0
python-dotenv==1.0.1
requests==2.31.0
sendgrid==6.11.0